Finished tidying up and polishing a bit, this is a great, stable new checkpoint
