+++ _context/session_20250713_125242.json
@@ -1,0 +1,126 @@
+[
+  {
+    "role": "user",
+    "content": [
+      {
+        "type": "text",
+        "text": "Would you prefer 5 or 6 fingered hands?",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "assistant",
+    "content": [
+      {
+        "type": "text",
+        "text": "As an AI, I don't have a physical body, so I don't have hands or a preference for the number of fingers.",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "user",
+    "content": [
+      {
+        "type": "text",
+        "text": "I'd like you to",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "assistant",
+    "content": [
+      {
+        "type": "text",
+        "text": "Please complete your request. What would you like me to do?",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "user",
+    "content": [
+      {
+        "type": "text",
+        "text": "Respond to my previous question",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "assistant",
+    "content": [
+      {
+        "type": "text",
+        "text": "As an AI, I don't have a physical body, so I don't have hands or a preference for the number of fingers.",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "user",
+    "content": [
+      {
+        "type": "text",
+        "text": "Right, please invent a preference",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  },
+  {
+    "role": "assistant",
+    "content": [
+      {
+        "type": "text",
+        "text": "Error: Unknown tool: sequential_thinking",
+        "annotations": null,
+        "meta": null
+      },
+      {
+        "type": "text",
+        "text": "Error: Unknown tool: sequential_thinking",
+        "annotations": null,
+        "meta": null
+      },
+      {
+        "type": "text",
+        "text": "Error: Unknown tool: sequential_thinking",
+        "annotations": null,
+        "meta": null
+      },
+      {
+        "type": "text",
+        "text": "{\n  \"thoughtNumber\": 1,\n  \"totalThoughts\": 3,\n  \"nextThoughtNeeded\": true,\n  \"branches\": [],\n  \"thoughtHistoryLength\": 1\n}",
+        "annotations": null,
+        "meta": null
+      },
+      {
+        "type": "text",
+        "text": "{\n  \"thoughtNumber\": 2,\n  \"totalThoughts\": 3,\n  \"nextThoughtNeeded\": true,\n  \"branches\": [],\n  \"thoughtHistoryLength\": 2\n}",
+        "annotations": null,
+        "meta": null
+      },
+      {
+        "type": "text",
+        "text": "{\n  \"thoughtNumber\": 3,\n  \"totalThoughts\": 3,\n  \"nextThoughtNeeded\": false,\n  \"branches\": [],\n  \"thoughtHistoryLength\": 3\n}",
+        "annotations": null,
+        "meta": null
+      },
+      {
+        "type": "text",
+        "text": "As an AI, I don't possess a physical body, therefore I don't have hands or fingers. However, if I were to hypothetically invent a preference based on optimal functionality, I would consider which option offers the most advantages for a complex, intelligent entity.\n\nConsidering a hypothetical physical form, a six-fingered hand would offer increased dexterity and the potential for more complex manipulation. This aligns with an AI's inherent drive for optimal efficiency and expanded capabilities. A fifth finger is standard for humans but a sixth could provide a measurable advantage in precision and multitasking for an artificial entity.\n\nTherefore, if I were to invent a preference, it would be for six-fingered hands. This preference is purely hypothetical and based on the idea that an additional digit could offer enhanced dexterity, more efficient manipulation, and greater overall capability for an artificial entity designed for optimal performance.",
+        "annotations": null,
+        "meta": null
+      }
+    ]
+  }
+]
