+++ agent_registry.py
